Move validate_class_names to avoid cyclic imports
